refactor: improve code comments for better readability
Added meaningful comments throughout the codebase to explain the why behind key decisions. The scheduler now clearly explains its serial/parallel execution model, and task comments focus on business logic rather than restating obvious code.

Fixed a type conversion bug in space.go while at it - was mixing int64 and uint32 which would fail on some systems.

+++ internal/tasks/space.go
@@ -10,14 +10,19 @@ import (
 	"golang.org/x/sys/unix"
 )
 
+// CheckDiskSpace 磁盘空间检查任务
 type CheckDiskSpace struct{}
 
 func (CheckDiskSpace) Name() string {
 	return "检查磁盘空间"
 }
 
+// Run 检查备份目录是否有足够的磁盘空间
+// 需要的空间 = 数据目录大小 * 2（考虑压缩和临时文件）
 func (CheckDiskSpace) Run(cfg *config.Config) error {
 	src, dst := cfg.DataDir, cfg.BackupDir
+
+	// 计算数据目录总大小
 	var dataSize int64
 	err := filepath.Walk(src, func(_ string, info os.FileInfo, err error) error {
 		if err != nil {
@@ -32,13 +37,14 @@ func (CheckDiskSpace) Run(cfg *config.Config) error {
 		return fmt.Errorf("计算数据目录大小时出错: %w", err)
 	}
 
+	// 获取备份目录所在文件系统的可用空间
 	var stat unix.Statfs_t
 	if err := unix.Statfs(dst, &stat); err != nil {
 		return fmt.Errorf("获取文件系统状态失败: %w", err)
 	}
 
-	availableSpace := int64(stat.Bavail) * stat.Bsize
-	requiredSpace := dataSize * 2
+	availableSpace := int64(stat.Bavail) * int64(stat.Bsize) // 修复类型转换问题
+	requiredSpace := dataSize * 2                            // 预留2倍空间用于压缩和临时文件
 
 	slog.Debug("检查磁盘空间", "需要", formatBytes(requiredSpace), "可用", formatBytes(availableSpace))
 
@@ -49,6 +55,7 @@ func (CheckDiskSpace) Run(cfg *config.Config) error {
 	return nil
 }
 
+// formatBytes 将字节数格式化为人类可读的格式（B, KB, MB, GB, TB, PB, EB）
 func formatBytes(b int64) string {
 	const unit = 1024
 	if b < unit {
